Task5 Spring boot handling Exception

+++ src/main/java/com/stackroute/service/MusicServiceImpl.java
@@ -1,5 +1,7 @@
 package com.stackroute.service;
 
+import com.stackroute.exceptions.MusicAlreadyExistsException;
+import com.stackroute.exceptions.TrackNotFoundException;
 import com.stackroute.respository.MusicRepository;
 import com.stackroute.domain.Music;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,13 +17,23 @@ public class MusicServiceImpl implements MusicService {
     public MusicServiceImpl(MusicRepository musicRepository)
     { this.musicRepository=musicRepository;}
     @Override
-    public Music saveTrack(Music track) {
-      return musicRepository.save(track);
+    public Music saveTrack(Music track)throws MusicAlreadyExistsException{
+        if(musicRepository.existsById(track.getTrackId()))
+        {
+            throw new MusicAlreadyExistsException("Already Exist");
+        }
+        track =  musicRepository.save(track);
+        if(track==null){
+            throw  new MusicAlreadyExistsException("Track Already Exists");
+        }
+
+      return track;
     }
 
     @Override
     public List<Music> getAllTracks() {
-        return musicRepository.findAll();
+
+        return  musicRepository.findAll();
 
     }
 
@@ -34,8 +46,18 @@ public class MusicServiceImpl implements MusicService {
     }
 
     @Override
-    public boolean deleteTrack(int trackId) {
+    public boolean deleteTrack(int trackId) throws TrackNotFoundException {
+        if((musicRepository.existsById(trackId)))
+        {
+            throw new TrackNotFoundException("Not Found");
+        }
         musicRepository.deleteById(trackId);
         return true;
+    }
+
+    @Override
+    public List<Music> findTitleByName(String trackName) {
+       List<Music> list= musicRepository.findTitleByName(trackName);
+        return  list;
     }
 }
